Patch solc for compatibility with Boost 1.69
Patches taken from upstream:

https://github.com/ethereum/solidity/pull/4772
https://github.com/ethereum/solidity/pull/5223
https://github.com/ethereum/solidity/pull/5226

It also requires a change to the arguments for the unit tests, since
boost changed how the options are parsed.

Update oldNixpkgs to the latest working version.

This revision has a `flake.nix` file so it no longer needs the input
config to specify `flake = false;`.

It also prints a warning about deprecated `stdenv.lib`.